Move fixtures to the top-level conftest.py to be shared with profiler tests
