+++ tests/conftest.py
@@ -50,6 +50,9 @@ SRC_DIR = os.path.join(PROJECT_DIR, "src")
 EXAMPLES_DIR = os.path.join(PROJECT_DIR, "examples")
 sys.path.append(SRC_DIR)
 
+if typing.TYPE_CHECKING:
+    from aiq.data_models.intermediate_step import IntermediateStep
+
 
 @pytest.fixture(name="test_data_dir")
 def test_data_dir_fixture():
@@ -351,3 +354,69 @@ def patched_async_memory_client(monkeypatch):
     mock_method = mock.MagicMock(return_value=None)
     monkeypatch.setattr(MemoryClient, "_validate_api_key", mock_method)
     return MemoryClient
+
+
+@pytest.fixture(name="rag_user_inputs")
+def rag_user_inputs_fixture() -> list[str]:
+    """Fixture providing multiple user inputs."""
+    return ["What is ML?", "What is NLP?"]
+
+
+@pytest.fixture(name="rag_generated_outputs")
+def rag_generated_outputs_fixture() -> list[str]:
+    """Fixture providing workflow generated outputs corresponding to user inputs."""
+    return ["ML is the abbreviation for Machine Learning", "NLP stands for Natural Language Processing"]
+
+
+@pytest.fixture(name="rag_intermediate_steps")
+def rag_intermediate_steps_fixture(rag_user_inputs, rag_generated_outputs) -> list[list["IntermediateStep"]]:
+    """
+    Fixture to generate separate lists of IntermediateStep objects for each user input.
+
+    Each list includes:
+    1. LLM_START, LLM_NEW_TOKENs, LLM_END
+    2. TOOL_START, and TOOL_END.
+
+    Returns:
+        (list for user_input_1, list for user_input_2)
+    """
+    from aiq.builder.framework_enum import LLMFrameworkEnum
+    from aiq.data_models.intermediate_step import IntermediateStep
+    from aiq.data_models.intermediate_step import IntermediateStepPayload
+    from aiq.data_models.intermediate_step import IntermediateStepType
+    from aiq.data_models.intermediate_step import StreamEventData
+
+    framework = LLMFrameworkEnum.LANGCHAIN
+    token_cnt = 10
+    llm_name = "mock_llm"
+    tool_name = "mock_tool"
+
+    def create_step(event_type, name=llm_name, input_data=None, output_data=None, chunk=None):
+        """Helper to create an `IntermediateStep`."""
+        return IntermediateStep(
+            payload=IntermediateStepPayload(event_type=event_type,
+                                            framework=framework,
+                                            name=name,
+                                            data=StreamEventData(input=input_data, output=output_data, chunk=chunk)))
+
+    step_lists = []  # Store separate lists
+
+    for user_input, generated_ouput in zip(rag_user_inputs, rag_generated_outputs):
+        tool_input = f"Get me the documents for {user_input}"
+        tool_output = f"Here is information I have on {user_input}"
+        generated_output = generated_ouput
+
+        steps = [
+            create_step(IntermediateStepType.LLM_START, input_data=user_input),
+            *[
+                create_step(IntermediateStepType.LLM_NEW_TOKEN, chunk=f"Token {i} for {user_input}")
+                for i in range(token_cnt)
+            ],
+            create_step(IntermediateStepType.LLM_END, input_data=user_input, output_data=generated_output),
+            create_step(IntermediateStepType.TOOL_START, name=tool_name, input_data=tool_input),
+            create_step(IntermediateStepType.TOOL_END, name=tool_name, input_data=tool_input, output_data=tool_output),
+        ]
+
+        step_lists.append(steps)  # Append separate list for each user input
+
+    return step_lists
